refactor config class into separate files
